feat(composables): remove useCustomRef
The original idea of useCustomRef was to provide the same features as a regular ref while allowing to easily customize getter and setter; it would provide default getter and setter so that, if not specified, default behavior is preserved.

However, to provide full ref support, the implementation becomes too complex: basically, a port of vue's ref function is required. The overhead comes from supporting reactive objects (if the ref's value is an object, it should be made reactive by default).

Thefore, useCustomRef becomes non-maintainable; vue's customRef should be used for very specific use cases (very well known ref values and how to interact with them).

To provide custom accessor methods, it'd be easier to override the ref's default accessor methods (see ref's implementation).