feat(utilities)!: remove getDayStart
Converting dates between timezones through an static timezone offset is not reliable if the offset is associated to a geographical location's timezone (which is usually — if not always — the case). A proper implementation should use IANA timezone identifiers in order to consider DTS or other historical timezone changes for a given location.